BUG/MINOR: h3: prevent hypothetical demux failure on int overflow
h3s stores the current demux frame type and length as a state info. It
should be big enough to store a QUIC variable-length integer which is
the maximum H3 frame type and size.

Without this patch, there is a risk of integer overflow if H3 frame size
is bigger than INT_MAX. This can typically causes demux state mismatch
and demux frame error. However, no occurence has been found yet of this
bug with the current implementation.

This should be backported up to 2.6.

(cherry picked from commit 35d9053b6832c419f9a94ff331b5c495df1cde9d)
(cherry picked from commit 1ef3723c229f9a18e7ba95e2000f477d12de7fc1)
[ad: remove h3 trace facility not supported on 2.6]

+++ src/h3.c
@@ -147,8 +147,8 @@ struct h3s {
 
 	enum h3s_t type;
 	enum h3s_st_req st_req; /* only used for request streams */
-	int demux_frame_len;
-	int demux_frame_type;
+	uint64_t demux_frame_len;
+	uint64_t demux_frame_type;
 
 	unsigned long long body_len; /* known request body length from content-length header if present */
 	unsigned long long data_len; /* total length of all parsed DATA */
